fixed an issue with my streak not updating
For some reason, my streak was not updating. Strangely enough, the issue was fixed by changing the number of query parameters I added to the streak's CDN. In this case, I added a gray color for the dates. If the issue arises again, try taking off the date parameter to see if the issue gets fixed (you can save the color of dates for later, in case the issue arises again). Super weird.